requirements.txt
pandas>=2.2.0
mysql-connector-python>=9.2.0
openpyxl>=3.1.5
python-dotenv>=1.0.1
mysql>=0.0.3
schedule>=1.2.2
matplotlib>=3.10.0